requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.3
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
